requirements.txt
ragas
langchain
langchain-openai
langchain-community
openai
python-dotenv
requests
pytest
pytest-asyncio
nltk
